+++ xmfcn-spring-cloud-wechat-service/src/main/java/com/cn/xmf/service/msg/dao/WxUserMessageDao.java
@@ -75,6 +75,6 @@ public interface WxUserMessageDao {
 	   * @param map
 	   * @return
 	   */
-	   public Integer  getTotalCount(Map map);
+	   public Long  getTotalCount(Map map);
 	   
 }
