+++ xmfcn-spring-cloud-wechat-service/src/main/java/com/cn/xmf/service/msg/service/WxUserMessageService.java
@@ -48,7 +48,7 @@ public class WxUserMessageService  {
           {
              return null;
           }
-          int totalcount =wxUserMessageHelperService.getTotalCount(json);
+          long totalcount =wxUserMessageHelperService.getTotalCount(json);
           List<WxUserMessage> list= null;
           if(totalcount>0)
           {
