+++ xmfcn-spring-cloud-wechat-service/src/main/java/com/cn/xmf/service/music/dao/MusicDao.java
@@ -75,6 +75,6 @@ public interface MusicDao {
 	   * @param map
 	   * @return
 	   */
-	   public Integer  getTotalCount(Map map);
+	   public Long  getTotalCount(Map map);
 	   
 }
