+++ xmfcn-spring-cloud-wechat-service/src/main/java/com/cn/xmf/service/user/dao/WxUserDao.java
@@ -68,6 +68,6 @@ public interface WxUserDao {
 	   * @param map
 	   * @return
 	   */
-	   public int  getTotalCount(Map map);
+	   public Long  getTotalCount(Map map);
 	   
 }
